Routed voxel_globe.quick.tasks for quick_vxl worker only
