+++ voxel_globe/tests/tasks.py
@@ -10,17 +10,6 @@ def success(self):
   return 123
 
 @shared_task(base=VipTask, bind=True)
-def vxl(self):
-  import boxm2_adaptor as ba
-  from StringIO import StringIO
-  from vsi.tools.redirect import Redirect
-  stdout = StringIO()
-  with Redirect(stdout_c=stdout):
-    ba.ocl_info()
-  stdout.seek(0)
-  return stdout.read()
-
-@shared_task(base=VipTask, bind=True)
 def python_crash(self):
   import time
   x = 15
@@ -42,8 +31,3 @@ def python_segfault(self):
 def run_ocl_info(self):
   import boxm2_adaptor as b
   b.ocl_info()
-
-@shared_task(base=VipTask, bind=True)
-def load_scene(self):
-  import boxm2_scene_adaptor as b
-  b.boxm2_scene_adaptor(r'D:\vip\tmp\tmptdnhd4\model\uscene.xml')
